Bugfix: Patroni falsely reported postgres as running
If postgres is not running due to the crash, panic or because it was
stopped by some external forcei, we need to change the internal state.
So far we always relied on state being set in the `Postgresql.start()`
method. Unfortunately not in all cases this method is reachable. For
example the `Postgresql.follow()` could raise an exception when calling
`write_recovery_conf()` and there is not enough disk space.

In order to solve it, we explicitly set the state when detected that the
postgres is not alive. In the `pause` we set it to `stopped`, otherwice
to `crashed` if it was `running` or `starting` before.