Make it possible to bypass kubernetes service
When running on K8s Patroni is communicating with API via the
`kubernetes` service, which is address is exposed via the
`KUBERNETES_SERVICE_HOST` environment variable. Like any other service
the `kubernetes` service is handled by kube-proxy, that depending on
configuration is either relying on userspace program or iptables for
traffic routing.
During K8s upgrade, when master nodes are replaced, it is possible that
kube-proxy doesn't update the service configuration in time and as a
result Patroni fails to update the leader lock and demotes postgres.

In order to improve the user experience and get more control on the
problem we make it possible to bypass the `kubernetes` service and
connect directly to API nodes.
The strategy is very simple:
1. Resolve list IPs of API nodes from the kubernetes endpoint
2. Stick to one of these IPs for API requests
3. Switch to a different IP if connected to IP not from the list
4. If the request fails, switch to another IP and retry

Such a strategy is already used for Etcd and proven to work quite well.
In order to enable it, you need either to set to `true`
`kubernetes.bypass_api_service` in the Patroni configuration file or
`PATRONI_KUBERNETES_BYPASS_API_SERVICE` environment variable.