Compatibility with PostgreSQL 13
So far Patroni was enforcing the same value of `wal_keep_segments` on
all nodes in the cluster. If the parameter was missing from the global
configuration it was using the default value 8.
In pg13 beta3 the `wal_keep_segments` was renamed to the `wal_keep_size`
and it broke Patroni.
If `wal_keep_segments` happened to be present in the configuration for
pg13, Paroni will recalculate the value to `wal_keep_size` assuming that
the `wal_segment_size` is 16MB. Sure, it is possible to get the real
value of `wal_segment_size` from pg_control, but since we are dealing
with the case of missconfiguration it is not worse time spend on it.