BUGFIX: pause on K8s
On K8s we don't remove the leader object but only update it.
For future work with the leader object we need to have its
resourseVersion and therefore we create Cluster.leader, which is not
empty (for other DCS Cluster.leader it is None).
The right way to identify missing leader is checking that the leader name is empty.