Advanced validation of PostgreSQL parameters
So far Patroni was performing a comparison of the old value (in the
pg_settings) with the new value (from Patroni configuration or from
DCS) in order to figure out if reload or restart are required when
the parameter has been changed. If the given parameter was missing
in the pg_settings, Patroni was ignoring it and not writing into the
config file.
In case if the Postgres is not running, no validation has been done
and parameters and values were written into the config as it is.

It is not a very common mistake, but people tend to mistype parameter
names or values.
Also, it happens that some parameters are removed in specific Postgres
versions and some new are added (e.g. checkpoint_segments replaced
with min_wal_size and max_wal_size in 9.5 or wal_keep_segments was
replaced with wal_keep_size in 13).

Writing nonexistent parameters or invalid values into the
postgresql.conf makes postgres unstartable. We can't solve this issue
100%, but at least we can approach this goal very close.