Fix Invalid os.symlink Calls when Moving Data Dir
Fixes the calls to 'os.symlink' within the 'move_data_directory'
function, as used to update the symlink(s) for any WAL or tablespace
directories following an init failure.  Specifically, the new name for
a WAL and/or tablespace directory is now passed in as in as the first
argument when calling os.symlink, while the second argument is now the
name of the symlink that is being recreated.  This aligns with the
Python documentation for os.symlink, which states the following
regarding the first two arguments for the os.symlink function
(arguments 'src' and 'dst'): "Create a symbolic link pointing to src
named dst" (https://docs.python.org/3/library/os.html#os.symlink).
With this change, all WAL and/or tablespaces directories will be
properly renamed following a init failure, and the "FileExistsError"
that would otherwise occur when attempting to recreate the associated
symlinks is no longer thrown.