Refactoring: rename last(_leader)_operation to the last_lsn
LSN is not really an operation, but Log Sequence Number.
The main reason for doing this change because we want to store cluster
optime not in the /optime/leader, but in the /status key, alongside with
confirmed_flush_lsn of permanent logical slots. This change requires a
lot of refactoring and will touch the same parts of source code, but if
we mix everything together it would be harder to perform a code-review.