Add compatibility with python-consul2
the good old python-consul is not maintained for a few years in a row,
therefore someone forked under a different name, but package files are
installed into the same location as for the old.

The API of both modules is mostly compatible therefore it wasn't hard
to add the support of both modules in Patroni.

Taking into account that python-consul is not a direct requirement for
Patroni, but extra, now the end user has a choice what to install.

Close https://github.com/zalando/patroni/issues/1810