Implemented the logical slots readiness check on replicas
logical slots copied to the replica are only safe to use when it is
known that the primary is holding the catalog_xid for them.

When it happened, Patroni writes to the log informational message that
the slot is safe to use after promotion.

If the replica was promoted when some of the logical slots are not
ready, Patroni wites the warning that some of permanent logical slots
where not ready and it might be unsafe using them.

P.S. One should be really unliky to catch the case when the table
structure was changed and autovacuum cleaned dead tuples in the
pg_catalog short before the failover.