Add support for encrypted TLS keys for REST API
The Python SSL library allows for the inclusion of a password
in its "load_cert_chain" function when setting up a SSLContext[1].
This allows for loading an encrypted key file in PEM representation
to be loaded into the certificate chain.

This commit adds the optional "keyfile_password" parameter to
the REST API block of configuration so that Patroni can load in
encrypted private keys when establishing its TLS socket.

This also adds the corollary "PATRONI_RESTAPI_KEYFILE_PASSWORD"
environmental variable, which has the same effect.

[1] https://docs.python.org/3/library/ssl.html#ssl.SSLContext.load_cert_chain