Add a new section to end on testing HA solution
Thanks to Ants Aasma for the suggestion and some tips on testing. I have this feeling this might be best in a separate document, but having a document called "testing" probably looks like code testing at first, so that's why I didn't go that direction. Plus, it makes the pull request easier this way.

It is intentionally non-exhaustive, but I am happy to add more if that comes back via review.