Write prev LSN to before checkpoint to optime if wal_achive=on
The #1527 introduced a feature of updating /optime/leader with the
location of the last checkpoint after the Postgres was shutdown cleanly.
If wal archiving is enabled, Postgres always switching the WAL file
before writing the checkpoint shutdown record. Normally it is not an
issue, but for clusters without too much write activity it could lead to
the situation that the visible replication lag becomes equal to the size
of a single WAL file. In fact, the previous file is mostly empty and
contains only a few records, and it should be safe to report the LSN of
the SWITCH record before the shutdown checkpoint.
In order to do that, Patroni first gets the output of the pg_controldata
and based on it two times calls pg_waldump. The first call reads the
checkpoint record (and verifies that this is really the shutdown
checkpoint). The next call reads the previous record and in case if it
is the 'xlog switch' (for 9.3 and 9.4) or 'SWITCH' (for 9.5+), the LSN
of the SWITCH record is written to the `/optime/leader`. In case of any
mismatch, failure to call pg_waldump or parse its output, the old
behavior is retained, i.e. `Latest checkpoint location` from the
pg_controldata is used.

Close https://github.com/zalando/patroni/issues/1860