Clear event after fetching leader optime
Commit 04b9fb9dd4c082a7a009e50e67641191e0ca9a5d introduced additional
conditions for updating cached version of the leader optime. It was
required for implementing health-checks based on replication lag in the
https://github.com/zalando/patroni/pull/1599

What in fact was forgotten, the event should be cleared after the new
value of the optime was fetched. Not doing so results in running the HA
loop more frequent than it is required.

In addition to that sligtly increase watch timeout, it will keep HA
loops in sync across all nodes in the cluster.

Close https://github.com/zalando/patroni/pull/1599