Don't watch for sync members
The watch for sync member(s) was introduced in order to give a signal
to the leader that one of the members set the nosync tag to true.
Since that time we have got a few more conditions that should be
notified about, therefore instead of watching for all members of the
cluster every cluster member checks whether the condition is met, and
instead of updating ZNode performs delete+create.

Since every member is already watching for new ZNodes to be created
inside the $scope/members/, they automatically get notified about
important changes, and therefore watching for sync members is redundant
and this commit removes it.