Implemented allowlist for REST API
If configured, only IPs that matching rules would be allowed to call
unsafe endpoints.
In addtition to that it is possible to automatically include IPs of
members of the cluster to the list.
If neither of the above is configured the old behavior is retained.

Partially address https://github.com/zalando/patroni/issues/1734