Correct log message for PG restart on replication config changes
When some replication-related changes happen (e.g., restore_command
is changed), Patroni restarts Postgres and the following can bee seen in logs:

INFO: changing primary_conninfo and restarting in progress   

This message is misleading because it mentions one particular parameter
but the change could happen with a different one (e.g., restore_command).

This is a quick improvement that makes the messaging better. In the future,
it would make sense to report the diff of changes (hence, particular root cause
of the restart) that is analyzed in config.py:record_missmatch.