Copy the logical slot over if advance failed due to the missing WAL
It could happen that the replica for some reason is missing the WAL file
required by the replication slot. The nature of this phenomenon is a bit
unclear, it might be that the WAL was recycled short before we copied
the slot file, but, we still need a solution to this problem.
If the `pg_replication_slot_advance()` fails with the `UndefinedFile`
exception (requested WAL segment pg_wal/... has already been removed),
the logical slot on the replica must be recreated.