Postpone writing postgresql.conf when joining running Postgres 12+
When joining already running Postgres, Patroni ensures that config
files are set according to expectations. With recovery parameters
converted to GUCs in Postgres 12 it became a little problem, because
when the `Postgresql` object is being created it is not yet known where
the given replica is supposed to stream from. It results in
postgresql.conf first being written without recovery parameters, and on
the next run of HA loop Patroni noticing inconsistencies and updating
the config one more time.

For Postgres v12 it is not a big issue, but for v13+ it resulted in
interruption of streaming replication.