Compatibility with kazoo-2.7+
Old versions of kazoo immediately discarded all requests to Zookeeper if
the connection is in the SUSPENDED state.  This is absolutely fine
because Patroni is handling retries on its own.
Starting from 2.7 kazoo started queueing requests instead of discarding
and as a result, the Patroni HA  loop was getting stuck until the
connection to Zookeeper is reestablished, causing no demote of the Postgres.
In order to return to the old behavior we override the `KazooClient._call()`
method.

In addition to that, we ensure that the
`Postgresql.reset_cluster_info_state()` method is called even if DCS failed
(the order of calls was changed in the #1820).

Close https://github.com/zalando/patroni/issues/1981