Fix problem with PyInstaller-frozen code and dot in the path
If PyInstaller-frozen application using patroni is placed in a
location that contains a dot in the path, the dcs_modules()
function in patroni.dcs breaks because pkgutil.iter_importers()
treats the given path as a package name.

Ref: pyinstaller/pyinstaller#5944

This can be avoided altogether by not passing a path to
iter_importers(), because PyInstaller's FrozenImporter is a
singleton and registered as top-level finder.