Improve compatibility with latest minor releases
The commit https://github.com/postgres/postgres/commit/93a0bf2390327a482ff37317f6e17547e735409e
changed the behavior of pg_settings.pending_restart. Mostly it will not
cause problems because usually people very rarely removing values from
the config, but one case is unique. If Patroni is restarted for an
upgrade and it finds that Postgres is up and running, it will rewrite
postgresql.conf and perform a reload. As a result, recovery parameters
will be removed from it for Postgres v12+ and the pending_restart flag
will be falsely set. In order to partially mitigate the problem before
it is fixed in Postgres we will skip recovery parameters when checking
pending_restart flags in the pg_settings.

In addition to that remove two postgres parameters from the validator
because they were reverted from Postgres v14.