Ensure pg_replication_slot_advance() doesn't timeout
The bigger gap between the slot flush lsn and the lsn we want to advance
to becomes more time it takes for the call to finish.
Once started failing the "lag" will grow more or less infinitely, that
have following negative side-effects:
1. Size of pg_wal on the replic will grow
2. Size of pg_wal on the primary will grow
3. Since the hot_standby_feedback is forcefully enabled, the primary
   will stop cleaning up dead tuples
I.e., we are not only in danger of running out of disk space, but also
increasing chances of transaction wraparound to happen.

Hence, when calling pg_replication_slot_advance() we want to set the
statement_timeout to 0.

Since the call is happening from the main HA loop and could take more
than `loop_wait`, the next heartbeat run could be delayed.
There is also a possibility that the call could take longer
than `ttl` and the member key/session in DCS for a given replica
expires.
The slot LSN in DCS is updated by the primary every `loop_wait` seconds.
Hence, we don't expect that the slot_advance() call will take
significantly longer than the `loop_wait` and therefore chances of the
member key/session to expire are very low.