Release the leader lock when pg_controldata reports "shut down"
Due to different reasons, it could happen that WAL archiving on the
primary stuck or significantly delayed. If we try to do a switchover
or shut it down, the shutdown will take forever and will not finish
until the whole backlog of WALs is processed.
In the meantime, Patroni keeps updating the leader lock, which prevents
other nodes from starting the leader race even if it is known that they
received/applied all changes.

The `Database cluster state:` is changed to "shut down" after:
- all data is fsynced to disk and the latest checkpoint is written to WAL
- all streaming replicas confirmed that they received all changes
(including the latest checkpoint)
- at the same time, the archiver process continues to do its job and
postmaster process is still running.

In order to solve this problem and make the switchover more
reliable/fast in a case when archive_command is slow/failing, Patroni
will remove the leader key immediately after pg_controldata started
reporting PGDATA as "shut down" cleanly and it verified that there is
at least one replica that received all changes. If there is no replicas
that fulfill the condition the leader key isn't removed and the old
behavior is retained, i.e. Patroni will keep updating it.