Use cached role as a fallback when postgres is slow
In some extreme cases Postgres could be so slow that the normal
monitoring query doesn't finish in a few seconds. It results in
the exception being raised from the
Postgresql._cluster_info_state_get() method, which could lead
to the situation that postgres isn't demoted on time.
In order to make it reliable we will catch the exception and use
the cached state of postgres (is_running and role) to determine
whether postgres is running as a primary.

Close https://github.com/zalando/patroni/issues/2073