Avoid unnecessary updates of the members ZNode.
When deciding whether the ZNode should be updated we rely on the cached
version of the cluster, which is updated only when members ZNodes are
deleted/created or the `/status`, `/sync`, `/failover`, `/config`,
or `/history` ZNodes are updated.

I.e. after the update of the current member ZNode succeeded the cache
become stale and all further updates are always performed even if the
value didn't change. In order to solve it we introduce the new attribute
in the Zookeeper class and will use it for memorizing the actual value
and for later comparison.