Handle missing timelines in history file when deciding to rewind
When restore_command is configured Postgres is trying to fetch/apply
all possible WAL segments and also fetch history files in order to
select the correct timeline. It could result in a situation that the new
history file will be missing some timelines.

Example:
- node1 demotes/crashes on timeline 1
- node2 promotes to timeline 2 and archives 00000002.history and crashes
- node1 recovers as a replica, "replays" 00000002.history and promotes
  to timeline 3

As a result, 00000003.history will not have the line with timeline 2,
because it never replayed any WAL segment from it.
The `pg_rewind` tool is supposed to correctly handle such case when
rewinding node2 from node1, but Patroni when deciding whether the rewind
should happen was searching for the exact timeline in the history file
from the new primary. The solution is to take the previous record from
the history file is the current replica timeline is missing.

Close https://github.com/zalando/patroni/issues/2118