Fix port in use error on certificate replacement
When switching certificates there is a race condition with a concurrent
API request. If there is one active during the replacement period then
the replacement will error out with a port in use error and Patroni gets
stuck in a state without an active API server.

Fix is to call server_close after shutdown which will wait for already
running requests to complete before returning.

Fixes #2184