Avoid running CHECKPOINT on remote master if credentials are missing
If pg_rewind is required to be run by a standby-cluster member, we need to make
sure the main cluster leader has persisted the timeline changes to disk, so we
were running CHECKPOINT on it, which requires a remote superuser connection. If
no superuser password was configured, this would fail (#2194).

Instead, only run CHECKPOINT (as superuser) if no rewind credentials were
provided; otherwise, use the rewind credentials to check if the correct
timeline is present in the main cluster leader's control file.