Handle rewind when demoted node was shut down
In case of DCS unavailability Patroni restarts Postgres in read-only.
It will cause pg_control to be updated with the
`Database cluster state: in archive recovery` and also could set
the `MinRecoveryPoint`.

When the next time Patroni is started it will assume that Postgres was
running as a replica and rewind isn't required and will try to start
the Postgres up. In this situation there is the chance that the start
will be aborted with the FATAL error message that looks like
`requested timeline 2 does not contain minimum recovery point 0/501E8B8 on timeline 1`.
On the next heart-beat Patroni will again notice that Postgres isn't
running which would lead to another start-fail attempt.
This loop is endless.

In order to mitigate the problem we do the following:
1. While figuring out whether the rewind is required we consider
   `in archive recovery` along with `shut down in recovery`.
2. If pg_rewind is required and the cluster state is
   `in archive recovery` we also perform recovery in
   a single-user mode.

Close https://github.com/zalando/patroni/issues/2242