Don't copy logical slot if there is mismatch with the config
A couple of times we have seen in the wild that the database for
the permanent logical slots was changed in the Patroni config.

It resulted in a below situation.
On the primary:
1. The slot must be dropped before creating it in a different DB.
2. Patroni fails to drop it because the slot is being in use.

Replica:
1. Patroni notice that the slot eixsts in the worng DB and
   successfully dropping the it.
2. Patroni copying the existing slot from the primary by it's name with
   Postgres restart.

And the loop repeats while the "wrong" slot exists on the primary.

Basically, replicas are continuosly restarting, what badly affects
availability.

In order to solve the problem, we will perform additional checks while
copying replication slot files from the primary and discard them if
`slot_type`, `database`, or `plugin` don't match our expectations.