Handle the TypeError exception
```
TypeError: '<' not supported between instances of 'NoneType' and 'int'
```

It could happen when the replica didn't started to stream using the
replication slot on the primary.

Also, it could be the symptom of a different issue:
`hot_standby_feedback` is diabled.

The last one we can check and show the annoying error in the log.