K8s: reset watchers if PATCH fails with 409
High CPU load on Etcd nodes and K8s API servers created a very strange
situation. A few clusters were running without a leader and the pod
which is ahead of others was failing to take a leader lock because
updates were failing with 409 (resource_version mismatch).

Effectively that means that TCP connections to K8s master nodes were
alive (in opposite case tcp keepalives would have resolved it), but no
UPDATE events were arriving via these connections.

The only good way to prevent this situation is intercepting 409 HTTP
responses and terminating existing TCP connections used for watches.

Now a few words about implementation. Unfortunately, watch threads are
waiting in the read() call most of the time and there is no good way to
interrupt them. But, the `socket.shutdown()` seems to do this job. We
already used this trick in the Etcd3 implementation.

This approach will help to mitigate the issue of not having the leader,
but at the same time replicas might still end up with the stale cluster
state cached and in the worst case will not stream from the leader.
Non-streaming replicas are less dangerous and could be covered by
monitoring and partially mitigated by correctly configured
`archive_command` and `restore_command`.